Separate file watching tests from normal tests
